requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
av==14.4.0
banks==2.1.2
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
docstring_parser==0.16
eval_type_backport==0.2.2
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-speech==2.32.0
google-cloud-texttospeech==2.27.0
google-genai==1.16.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.2
griffe==1.7.3
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.0
livekit==1.0.8
livekit-agents==1.0.22
livekit-api==1.0.2
livekit-plugins-cartesia==1.0.22
livekit-plugins-deepgram==1.0.22
livekit-plugins-google==1.0.22
livekit-plugins-noise-cancellation==0.2.4
livekit-plugins-silero==1.0.22
livekit-protocol==1.0.3
llama-cloud==0.1.19
llama-cloud-services==0.6.22
llama-index==0.12.37
llama-index-agent-openai==0.4.7
llama-index-cli==0.4.1
llama-index-core==0.12.37
llama-index-embeddings-gemini==0.3.2
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-llms-openai==0.3.42
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-qdrant==0.6.0
llama-parse==0.6.22
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.6
onnxruntime==1.22.0
openai==1.79.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
portalocker==2.10.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
pypdf==5.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.14.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
SQLAlchemy==2.0.41
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
types-protobuf==4.25.0.20240417
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
watchfiles==1.0.5
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
